1. 整合组件 2. 封装重复代码
